reworking homeview UI
